Reorganize all the things
